chore: register form request

+++ app/Http/Requests/Api/Auth/RegisterRequest.php
@@ -16,12 +16,12 @@ class RegisterRequest extends ApiRequest
     {
         return [
             'name' => ['required', 'string', 'max:255'],
-            'age' => ['required', 'numeric', 'min:1', 'max:150'],
+            'age' => ['required', 'integer', 'min:1', 'max:150'],
             'email' => ['required', 'email:dns', 'max:255', 'unique:users,email'],
             'phone' => ['required', 'string', 'max:255', 'unique:users,phone'],
             'gender' => ['required', 'string', 'max:255', new GenderRule],
             'address' => ['required', 'string'],
-            'grade_id' => ['required', 'numeric', 'min:1', 'exists:grades,id'],
+            'grade_id' => ['required', 'integer', 'min:1', 'exists:grades,id'],
             'password' => ['required', 'string', 'min:8', 'max:255', 'confirmed']
         ];
     }
